requirements.txt
amadeus==3.2.0
PyYAML==5.1.2
pprint==0.1
requests==2.22.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
idna==2.8
PyQt5==5.13.0
PyQt5-sip==12.7.0
python-dateutil==2.8.1
python-dotenv==0.10.3
six==1.13.0
urllib3==1.25.7
googlemaps==v3.1.4
PyQtWebEngine==5.13.2
pytest==5.3.2